+++ src/Console/Command/Status.php
@@ -2,6 +2,7 @@
 
 namespace Sokil\Mongo\Migrator\Console\Command;
 
+use Sokil\Mongo\Exception;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -19,11 +20,35 @@ class Status extends \Sokil\Mongo\Migrator\Console\Command
                 InputOption::VALUE_OPTIONAL,
                 'Environment name'
             )
+            ->addOption(
+                '--length',
+                '-l',
+                InputOption::VALUE_OPTIONAL,
+                'Limit list by number of last revisions. If not set, show all revisions.'
+            )
             ->setHelp('Show list of migrations with status of applying');
     }
-    
+
+    /**
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     *
+     * @return int|null|void
+     *
+     * @throws Exception
+     */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        // length of list
+        $length = $input->getOption('length');
+        if (empty($length)) {
+            $length = null;
+        } elseif (is_numeric($length)) {
+            $length = (int)$length;
+        } else {
+            throw new \InvalidArgumentException('Length must be numeric, if specified');
+        }
+
         // environment
         $environment = $input->getOption('environment');
         if (!$environment) {
@@ -45,7 +70,7 @@ class Status extends \Sokil\Mongo\Migrator\Console\Command
         // body
         $manager = $this->getManager();
         
-        foreach ($manager->getAvailableRevisions() as $revision) {
+        foreach ($manager->getAvailableRevisions($length) as $revision) {
             if ($manager->isRevisionApplied($revision->getId(), $environment)) {
                 $status = '<info>up</info>' . str_repeat(' ', $columnWidth[1] - 2);
             } else {